+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 ```
+* Commit: 762fd10 - v111
+| Author: fjrodafo - Thu, 28 Nov 2024 19:32:04 +0000
+| 
 * Commit: 77f6c5e - v110
 | Author: fjrodafo - Thu, 28 Nov 2024 19:23:56 +0000
 |